# docs/api/openapi.phase0.yaml
openapi: 3.0.3
info:
  title: Bizboard API
  version: 0.1.0-phase0
  description: Initial API contract for Bizboard MVP.
servers:
  - url: https://api.bizboard.test/api/v1
    description: Local/Staging API
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      tags: [Catalog]
      security: []
      summary: List product/business categories
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [product, business, service, content] }
        - in: query
          name: parentId
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
  /products:
    get:
      tags: [Catalog]
      security: []
      summary: Search/list products
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: categorySlug
          schema: { type: string }
        - in: query
          name: brandSlug
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        '200':
          description: Product search result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductListResponse' }
  /products/{slug}:
    get:
      tags: [Catalog]
      security: []
      summary: Get product by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
...
  /payments/{orderId}/attempts:
    post:
      tags: [Payments]
      summary: Create payment attempt and gateway redirect URL
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gateway]
              properties:
                gateway: { type: string, enum: [zarinpal, zibal] }
      responses:
        '201':
          description: Payment attempt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentAttempt' }
  /payments/callback/{gateway}:
    post:
      tags: [Payments]
      security: []
      summary: Gateway callback endpoint. Must verify server-to-server.
      parameters:
        - in: path
          name: gateway
          required: true
          schema: { type: string, enum: [zarinpal, zibal] }
      responses:
        '200': { description: Callback processed idempotently }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        parentId: { type: string, format: uuid, nullable: true }
        slug: { type: string }
        title: { type: string }
        categoryType: { type: string }
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProductCard' }
        nextCursor: { type: string, nullable: true }
    ProductCard:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        imageUrl: { type: string, nullable: true }
        minPrice: { type: number, nullable: true }
        maxPrice: { type: number, nullable: true }
        offerCount: { type: integer }
        ratingAvg: { type: number }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductCard'
        - type: object
          properties:
            description: { type: string, nullable: true }
            aiSummary: { type: string, nullable: true }
            specs: { type: object }
            offers:
              type: array
              items: { $ref: '#/components/schemas/ProductOffer' }
    ProductOffer:
      type: object
      properties:
        id: { type: string, format: uuid }
        merchantId: { type: string, format: uuid }
        price: { type: number }
        oldPrice: { type: number, nullable: true }
        currency: { type: string }
        availability: { type: string }
        internalCheckoutEnabled: { type: boolean }
        merchantProductUrl: { type: string, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
...
    PaymentAttempt:
      type: object
      properties:
        id: { type: string, format: uuid }
        gateway: { type: string }
        status: { type: string }
        redirectUrl: { type: string }